+++ release_notes.md
@@ -2,4 +2,7 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR/#issue)
 -->
+
+- Add Retry Policy Attributes (#977, #971)
 - Bug fix - GetOutputBindings returns incorrect data when OutputBindingData is not set (#983)
+